backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "beanie>=1.29.0",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.19.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.1.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "sse-starlette>=1.8.2",
]
